EigenMatToPointCloud2 - Use rows() of Eigen::MatrixX3f &points
The usage of points.size() in EigenMatToPointCloud2 to
CreatePointCloudMsg() was incorrect, it should have been points.rows().

Using points.size() leads 3x more points being allocated in the
point cloud than necessary which bloats the message slowing things down.

These extra points are at the origin of the pointcloud.

This can manifest as an artificial hump directly under
the origin of the point cloud.

Signed-off-by: Mike Wake <macwake@gmail.com>